requirements.txt
aiosqlite==0.17.0
anyio==3.6.2
black==23.3.0
click==8.1.3
fastapi==0.95.0
h11==0.14.0
hashids==1.3.1
idna==3.4
iso8601==1.1.0
makefun==1.15.1
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
pydantic==1.10.7
PyJWT==2.6.0
pypika-tortoise==0.1.6
python-dotenv==1.0.0
pytz==2023.3
sniffio==1.3.0
starlette==0.26.1
tomli==2.0.1
tortoise-orm==0.19.3
typing_extensions==4.5.0
uvicorn==0.21.1
